refactor(gui): reorganize misc UI modules into common/dev/tx_history
- Move shared widgets (amount field, confirm window, chain/wallet selects) out of `misc` into a dedicated `common` module; `AmountFieldWithCurrencySelect` is now `AmountField`
- Promote `misc::dev` to its own top-level `dev` module
- Promote `misc::tx_history` to its own top-level `tx_history` module
- Drop the catch-all `misc` module from `ui/mod.rs`; update all imports across `gui/mod.rs`, `dapps/across`, `dapps/uniswap/swap`, and `send_crypto` to the new paths
