+++ src/gui/ui/common/mod.rs
@@ -1,12 +1,11 @@
+//! Common UI components
+
 pub mod amount_field;
 pub mod chain_select;
-pub mod dev;
-pub mod sync;
-pub mod tx_history;
 pub mod wallet_select;
 pub mod windows;
 
-pub use amount_field::AmountFieldWithCurrencySelect;
+pub use amount_field::AmountField;
 pub use chain_select::ChainSelect;
 pub use wallet_select::WalletSelect;
 pub use windows::{ConfirmWindow, LoadingWindow, MsgWindow, UpdateWindow};
